Clean up awkawk a bit
This cleans up some of this so I can shove it into a container (awkawk
was written before I'd touched any containerization stuff). There's
a (ostensibly) graceful shutdown, glog is used instead of std log (that
one is personal preference, but the more important point is that it
logs the listen address), and there's a minor change to some of the
responses to prevent excessively large responses.

The behavior of the server changes a little bit. You can no longer pass
multiple form fields of the same name and have them concatenated, so
it's first-wins. Errors are reported differently as well (and encoded
slightly more correctly).